fix: removed shared logic, this avoid bundling issues
